requirements.txt
websocket-client
python-osc
PyQt5
numpy
matplotlib